вывел поинтер, ура

+++ ft_parser.c
@@ -34,7 +34,7 @@ void	ft_pwtype(t_struct *box)
 		if (box->argv1[i] == 'X')
 			ft_putnbr_bx(va_arg(box->argument_pointer, int));
 		if (box->argv1[i] == 'p')
-			ft_putnbr_p(va_arg(box->argument_pointer, int));
+			ft_putnbr_p(va_arg(box->argument_pointer, unsigned long));
 		if (box->argv1[i - 1] == '%')
 			i++;
 		if (box->argv1[i] == '\n')
